+++ tests/test_config.py
@@ -8,6 +8,19 @@ from visual_graph_datasets.config import load_config
 from visual_graph_datasets.config import Config
 from visual_graph_datasets.util import TEMPLATE_ENV
 
+from .util import TestingConfig
+
+
+def test_testing_config():
+    """
+    TestingConfig is a utility context manager which can be used to create isolated environments for the
+    testing purposes.
+    """
+    with TestingConfig() as config:
+        assert isinstance(config, Config)
+        assert isinstance(config.get_datasets_path(), str)
+        assert os.path.exists(config.get_datasets_path())
+
 
 def test_load_config_works_when_no_file_exists():
     config_dict = load_config('')
@@ -73,5 +86,5 @@ def test_config_retrieve_nested_value():
         config = Config()
         config.load(config_path)
 
-        value = config.retrieve_nested_with_default('base/provider', '')
+        value = config.retrieve_nested_with_default('providers/main/type', '')
         assert value != ''