+++ visual_graph_datasets/processing/base.py
@@ -1,0 +1,9 @@
+import typing as t
+
+
+def identity(value: t.Any) -> t.Any:
+    return value
+
+
+def list_identity(value: t.Any, dtype: type = float) -> t.Any:
+    return [dtype(value)]